Create README.md
This is the link for the project files: https://drive.google.com/file/d/1imTxnVqnSffgYhuyioWwtpwcPfKe8dIs/view?usp=sharing

Hello! 
This is the implementation of paper "FedAGAT: Real-time traffic flow prediction based on federated community and adaptive graph attention network", so if you found this code useful kindly cite the paper. The implementation has gone through several updates, so while this may not be the finalized release version, it reflects the latest version on my side. I hope it will be helpful for your work.
Thank you